requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pandas==2.1.3
python-dotenv==1.0.0
colorama==0.4.6
